==> notebooks/01_kenlm_prepare_dataset.ipynb <==
# from huggingface_hub import hf_hub_download
from datasets import load_dataset
dataset = load_dataset('thexdk/wiki_ru', split="train", streaming=True)

import ftfy
from razdel import tokenize
from tqdm import tqdm

# total_document = [None for _ in range(1e6)]
with open('wiki_ru', 'a', encoding="utf-8") as f:
    for i in tqdm(dataset, total=1e6):
        text = i['text']
        if not text:
            continue # just in case
        res = ' '.join([j.text.lower() for j in tokenize(text)])
        res = ' '.join([
            '<space>' if j == ' ' else j 
            for j in res
        ])
        print(res, file=f, end=' ')

dataset = load_dataset('daje/en_wiki', split="train", streaming=True)

c = 0
with open('wiki_en', 'a', encoding="utf-8") as f:
    for i in tqdm(dataset, total=1e6):
        text = i['text']
        # text = ftfy.fix_text(text, normalization='NFKD')
        if not text:
            continue # just in case
        res = ' '.join([j.text.lower() for j in tokenize(text)])
        res = ' '.join([
            '<space>' if j == ' ' else j 
            for j in res
        ])
        print(res, file=f, end=' ')
        c += 1
        if c == 1000000:
            break

# [j for j in ' '.join([j.text.lower() for j in tokenize(text)])][5]
i

with open('wiki_ru', 'a') as f:
    print(res, file=f, end=' ')

# dataset = load_dataset('daje/en_wiki')

# !pip install razdel

# dataset['train']['text']

# dataset['train']['text'][0]
