==> notebooks/02_kenlm_train.ipynb <==
!head -c 1000 drive/MyDrive/Colab\ Notebooks/wiki_first1m_chars

!sudo apt-get install build-essential libboost-all-dev cmake zlib1g-dev libbz2-dev liblzma-dev

!wget -O - https://kheafield.com/code/kenlm.tar.gz |tar xz

!mkdir kenlm/build && cd kenlm/build

!cd kenlm/build/ && cmake ..

!cd kenlm/build/ && make -j 2

!./kenlm/build/bin/lmplz -o 1 -T /tmp/ -S 80% < drive/MyDrive/Colab\ Notebooks/wiki_first1m_chars > char_5gram_ru_en.arpa

!./kenlm/build/bin/lmplz -o 5 -T /tmp/ -S 80% --vocab_estimate 100000 < drive/MyDrive/wiki_ru_en_1m.gz > char_5gram_ru_en.arpa
# !./kenlm/build/bin/lmplz -o 7 -T /tmp/ -S 80% --vocab_estimate 100000 < drive/MyDrive/wiki_ru_en_1m.gz > char_7gram_ru_en.arpa

!tail -n 100 char_5gram_ru_en.arpa

!ls -al

!./kenlm/build/bin/build_binary -s char_5gram_ru_en.arpa char_5gram_ru_en.bin
