==> notebooks/03_kenlm_fix_texts.ipynb <==
import kenlm

model = kenlm.Model('../../kenlm/char_5gram_ru_en.bin')

def tok(t):
    return ' '.join(['<space>' if j == ' ' else j for j in t])

import pandas as pd

df = pd.read_parquet('name_desc_avito.parquet')

import re

df_name_eng = df['name'][df['name'].str.match('[a-zA-Z]')]
df_name_eng

df_desc_eng = df['description'][df['description'].str.match('[a-zA-Z]')]
df_desc_eng

# from transformers import BertTokenizer, BertModel
# tokenizer = BertTokenizer.from_pretrained('bert-base-multilingual-uncased')

# tokenizer('Sakura SASP377XE Кapтpидж cовмеcтимый Сaкуpa SP-377XE чеpный', return_length=True)['length']

def color_eng(text):
    out = []
    for c in text:
        if c.lower() in 'qwertyuiopasdfghjklzxcvbnm':
            out.append('\x1b[0;30;41m' + c + '\x1b[0m')
        else:
            out.append(c)
    print(''.join(out))
    
color_eng(df_desc_eng.iloc[0])

MAPPING_EN_RU = {
    'e': 'е',
    'y': 'у',
    'o': 'о',
    'p': 'р',
    'a': 'а',
    'k': 'к',
    'x': 'х',
    'c': 'с',
    'E': 'Е',
    'T': 'Т',
    'O': 'О',
    'P': 'Р',
    'A': 'А',
    'H': 'Н',
    'K': 'К',
    'X': 'Х',
    'C': 'С',
    'B': 'В',
    'M': 'М',
}
MAPPING_RU_EN = {v: k for k, v in MAPPING_EN_RU.items()}

import kenlm

model = kenlm.Model('../../kenlm/char_5gram_ru_en.bin')

import re

has_eng = re.compile('[a-zA-Z]')
r1 = re.compile(r'\b[a-zA-Zа-яА-ЯёЁ]+\b')
r2 = re.compile(r'[a-zA-Z]+')
r3 = re.compile(r'[а-яА-ЯёЁ]+')

def fix_text_shit(
    text, 
    lookaround=2, 
    # pad_thr=4, 
    skip_less_than=3,
):
    if not has_eng.match(text):
        return text
    ret_text = [i for i in text]
    for match in r1.finditer(text):
        # print(match, text[match.span()[1]:match.span()[0]])
        # print(re.fullmatch('[а-яА-ЯёЁ]+', text[match.span()[1]:match.span()[0]]))
        if match.span()[1] - match.span()[0] < skip_less_than:
            continue # pass if word too small
        if r2.fullmatch(text[match.span()[0]:match.span()[1]]):
            continue # pass if only eng
        if r3.fullmatch(text[match.span()[0]:match.span()[1]]):
            continue # pass if only ru
        slice_ = slice(match.span()[0] - lookaround, match.span()[1] + lookaround)
        # slice_ = slice(match.span()[0] - lookaround, match.span()[1] + lookaround) if match.span()[1] - match.span()[0] > pad_thr else slice(match.span()[0] - lookaround - 1, match.span()[1] + lookaround + 1)
        splitted = [i for i in text[slice_].lower()]
        splitted_ru = ''.join([MAPPING_EN_RU.get(i, i) for i in splitted])
        splitted_en = ''.join([MAPPING_RU_EN.get(i, i) for i in splitted])
        ppl_ru = model.perplexity(tok(splitted_ru))
        ppl_en = model.perplexity(tok(splitted_en))
        # ppl_ru = ppl(splitted_ru)
        # ppl_en = ppl(splitted_en)
        # print(splitted_ru, 'vs', splitted_en)
        if ppl_ru < ppl_en:
            # print('-- conflict resolved towards ru', match, ppl_ru, ppl_en)
            ret_text[match.span()[0]:match.span()[1]] = [MAPPING_EN_RU.get(i, i) for i in text[match.span()[0]:match.span()[1]]]
        elif ppl_ru > ppl_en:
            # print('-- conflict resolved towards en', match, ppl_ru, ppl_en)
            ret_text[match.span()[0]:match.span()[1]] = [MAPPING_RU_EN.get(i, i) for i in text[match.span()[0]:match.span()[1]]]
        else:
            # print('-- leave as is', match, ppl_ru, ppl_en)
            continue # leave as is
        
    res = ''.join(ret_text)
    res = res.replace('нa', 'на').replace(' c ', ' с ').replace(
        ' cо ', ' со ').replace(' сo ', ' со ').replace(
        ' co ', ' со ').replace('вo', 'во').replace(
        ' кo ', ' ко ').replace(' o ', ' о ').replace(
        'oб', 'об').replace('oт', 'от').replace(
        'зa', 'за').replace('пo', 'по').replace(
        'дo', 'до').replace(' y ', ' у ').replace(
        'cм', 'см').replace('гp', 'гр').replace(
        'пpo', 'про').replace('дa', 'да').replace(
        'нe', 'не').replace('тo', 'то').replace(
        'жe', 'же').replace('pyб', 'руб').replace(
        'eд', 'ед').replace('oна', 'она').replace(
        'онa', 'она').replace('oнa', 'она').replace(
        'oн', 'он').replace('eго', 'его').replace(
        'егo', 'его').replace('eгo', 'его').replace(
        'ниx', 'них').replace('иx', 'их').replace(
        'вcе', 'все').replace('всe', 'все').replace('вce', 'все')
        
    return res

test_text = df_desc_eng.iloc[4]
color_eng(test_text)

color_eng(fix_text_shit(test_text, lookaround=2, skip_less_than=3))

%timeit fix_text_shit(test_text)

sum(df_desc_eng_fixed != df['description'])

sum(df_name_eng_fixed != df['name'])

from tqdm import tqdm
tqdm.pandas()

# def fix(x): 
#     return fix_text_shit(x) if has_eng.match(x) else x

df_desc_eng_fixed = df['description'].progress_apply(fix_text_shit)
df_name_eng_fixed = df['name'].progress_apply(fix_text_shit)

df['description_fixed'] = df_desc_eng_fixed
df['name_fixed'] = df_name_eng_fixed

df = df[['name_fixed', 'description_fixed']]
df

df.to_parquet('names_descs_avito_fixed.parquet')
