==> notebooks/06_get_texts_dfs.ipynb <==
import pandas as pd

df = pd.read_parquet('../data/preprocessed/all_products_preprocessed.parquet')
df_train = df[df['is_double'] != -1]
df_test = df[df['is_double'] != 1]

cols_to_use = [
    'variantid_1', 'variantid_2',
    'name_1', 'name_2',
    'description_1', 'description_2',
    'category_level_1_1', 'category_level_1_2',
    'category_level_2_1', 'category_level_2_2',
    'category_level_3_1', 'category_level_3_2',
    'category_level_4_1', 'category_level_4_2',
    'characteristic_attributes_mapping_1', 'characteristic_attributes_mapping_2',
]

df_train = df_train[cols_to_use + ['group_id', 'is_double']]
df_test = df_test[cols_to_use]

df_train.to_parquet('../data/preprocessed/train_texts.parquet')
df_test.to_parquet('../data/preprocessed/test_texts.parquet')
