==> notebooks/08_prepare_data_for_embedders.ipynb <==
import pandas as pd
import gc

df = pd.read_parquet('../data/preprocessed/train_texts.parquet')

df = df.sort_values(by=['variantid_1', 'variantid_2'])
df = df.sample(len(df), random_state=42)

cat_cols = [
    'category_level_1_1', 'category_level_1_2', 
    'category_level_2_1', 'category_level_2_2',
    'category_level_3_1', 'category_level_3_2', 
    'category_level_4_1', 'category_level_4_2'
]

for col in cat_cols:
    df[col] = df[col].fillna('none')

df['product_a'] = 'NAME: ' + df['name_1'] + '\nCATEGORY: ' + df['category_level_1_1'] + '->' + df['category_level_2_1'] + '->' + df['category_level_3_1'] + '->' + df['category_level_4_1'] + '\nDESCRIPTION: ' + df['description_1']
df['product_b'] = 'NAME: ' + df['name_2'] + '\nCATEGORY: ' + df['category_level_1_2'] + '->' + df['category_level_2_2'] + '->' + df['category_level_3_2'] + '->' + df['category_level_4_2'] + '\nDESCRIPTION: ' + df['description_2']

del df['name_1'], df['name_2'], df['category_level_1_1'], df['category_level_1_2'], df['category_level_2_1'], df['category_level_2_2'], df['category_level_3_1'], df['category_level_3_2'], df['category_level_4_1'], df['category_level_4_2'], df['description_1'], df['description_2']
gc.collect()

df = df[['variantid_1', 'variantid_2', 'product_a', 'product_b', 'group_id', 'is_double']]

df.rename(columns={
    'product_a': 'sentence1',
    'product_b': 'sentence2',
    'is_double': 'label',
}, inplace=True)

df.to_parquet('avito-for-dl-training.parquet')

df = pd.read_parquet('../data/preprocessed/test_texts.parquet')

df = df.sort_values(by=['variantid_1', 'variantid_2'])

cat_cols = [
    'category_level_1_1', 'category_level_1_2', 
    'category_level_2_1', 'category_level_2_2',
    'category_level_3_1', 'category_level_3_2', 
    'category_level_4_1', 'category_level_4_2'
]

for col in cat_cols:
    df[col] = df[col].fillna('none')

df['product_a'] = 'NAME: ' + df['name_1'] + '\nCATEGORY: ' + df['category_level_1_1'] + '->' + df['category_level_2_1'] + '->' + df['category_level_3_1'] + '->' + df['category_level_4_1'] + '\nDESCRIPTION: ' + df['description_1']
df['product_b'] = 'NAME: ' + df['name_2'] + '\nCATEGORY: ' + df['category_level_1_2'] + '->' + df['category_level_2_2'] + '->' + df['category_level_3_2'] + '->' + df['category_level_4_2'] + '\nDESCRIPTION: ' + df['description_2']

del df['name_1'], df['name_2'], df['category_level_1_1'], df['category_level_1_2'], df['category_level_2_1'], df['category_level_2_2'], df['category_level_3_1'], df['category_level_3_2'], df['category_level_4_1'], df['category_level_4_2'], df['description_1'], df['description_2']
gc.collect()

df = df[['variantid_1', 'variantid_2', 'product_a', 'product_b']]

df.rename(columns={
    'product_a': 'sentence1',
    'product_b': 'sentence2',
}, inplace=True)

df.to_parquet('avito-for-dl-testing.parquet')
