==> notebooks/09_prepare_data_for_clfs.ipynb <==
import pandas as pd
import gc

DESC_N = 1000

df1 = pd.read_parquet('../data/preprocessed/train_texts.parquet')
df1['description_1'] = df1['description_1'].apply(lambda x: x[:DESC_N])
df1['description_2'] = df1['description_2'].apply(lambda x: x[:DESC_N])

cat_cols = [
    'category_level_1_1', 'category_level_1_2', 
    'category_level_2_1', 'category_level_2_2',
    'category_level_3_1', 'category_level_3_2', 
    'category_level_4_1', 'category_level_4_2'
]

for col in cat_cols:
    df1[col] = df1[col].fillna('none')

df1['product_row'] = 'NAME1: ' + df1['name_1'] + '; NAME2: ' + df1['name_2'] + '\nCATEGORY1: ' + df1['category_level_1_1'] + '->' + df1['category_level_2_1'] + '->' + df1['category_level_3_1'] + '->' + df1['category_level_4_1'] + '; CATEGORY2: ' + df1['category_level_1_2'] + '->' + df1['category_level_2_2'] + '->' + df1['category_level_3_2'] + '->' + df1['category_level_4_2'] + '\nDESCRIPTION1: ' + df1['description_1'] + '; DESCRIPTION2: ' + df1['description_2']

gc.collect()

df1 = df1[['variantid_1', 'variantid_2', 'product_row', 'group_id', 'is_double']]

df1.to_parquet('avito-for-dl-train-rev.parquet')

df1 = pd.read_parquet('../data/preprocessed/test_texts.parquet')
df1['description_1'] = df1['description_1'].apply(lambda x: x[:DESC_N])
df1['description_2'] = df1['description_2'].apply(lambda x: x[:DESC_N])

cat_cols = [
    'category_level_1_1', 'category_level_1_2', 
    'category_level_2_1', 'category_level_2_2',
    'category_level_3_1', 'category_level_3_2', 
    'category_level_4_1', 'category_level_4_2'
]

for col in cat_cols:
    df1[col] = df1[col].fillna('none')

df1['product_row'] = 'NAME1: ' + df1['name_1'] + '; NAME2: ' + df1['name_2'] + '\nCATEGORY1: ' + df1['category_level_1_1'] + '->' + df1['category_level_2_1'] + '->' + df1['category_level_3_1'] + '->' + df1['category_level_4_1'] + '; CATEGORY2: ' + df1['category_level_1_2'] + '->' + df1['category_level_2_2'] + '->' + df1['category_level_3_2'] + '->' + df1['category_level_4_2'] + '\nDESCRIPTION1: ' + df1['description_1'] + '; DESCRIPTION2: ' + df1['description_2']

gc.collect()

df1 = df1[['variantid_1', 'variantid_2', 'product_row']]

df1.to_parquet('avito-for-dl-test-rev.parquet')
