==> notebooks/17_get_pretrain_e5large_cossims_train.ipynb <==
import gc
import torch
import pandas as pd
import numpy as np
from tqdm.auto import tqdm
from sentence_transformers import SentenceTransformer
from concurrent.futures import ThreadPoolExecutor

train_df = pd.read_parquet('avito-for-dl-train.parquet')

model1 = SentenceTransformer("intfloat/multilingual-e5-large", device='cuda:0') # deepvk/USER-bge-m3
model2 = SentenceTransformer("intfloat/multilingual-e5-large", device='cuda:1') # deepvk/USER-bge-m3

batch_size = 312
cossims = []

def encode_on_device(model, sentences, device):
    return model.encode(
        sentences,
        batch_size=batch_size,
        normalize_embeddings=True,
        convert_to_numpy=True,
        show_progress_bar=False,
        device=device
    )

for i in tqdm(range(0, len(dftrain_df), batch_size)):
    batch = train_df.iloc[i:i+batch_size]
    
    sentences1 = batch['sentence1'].tolist()
    sentences2 = batch['sentence2'].tolist()
    
    with ThreadPoolExecutor(max_workers=2) as executor:
        future_embs1 = executor.submit(encode_on_device, model1, sentences1, torch.device('cuda:0'))
        future_embs2 = executor.submit(encode_on_device, model2, sentences2, torch.device('cuda:1'))
        
        embs1 = future_embs1.result()
        embs2 = future_embs2.result()
    
    batch_cossims = np.einsum('ij,ij->i', embs1, embs2)
    cossims.extend(batch_cossims.tolist())
    
    del embs1, embs2, batch
    gc.collect()

train_df['e5large_cossim'] = cossims

train_df[['variantid_1', 'variantid_2', 'e5large_cossim']].to_parquet('e5large_cossims_train.parquet')
