==> notebooks/25_get_trained_rubert_base_preds_rev.ipynb <==
import os
import pandas as pd

from sklearn.model_selection import StratifiedGroupKFold
from sklearn.metrics import precision_recall_curve, auc

from dataclasses import dataclass
from datasets import Dataset

import torch
import torch.nn as nn
import torch.nn.functional as F

from transformers import (
    AutoModelForSequenceClassification,
    AutoTokenizer,
    EvalPrediction,
    Trainer,
    TrainingArguments,
    DataCollatorWithPadding,
)

import numpy as np
import _codecs

torch.serialization.add_safe_globals([
    np.core.multiarray.scalar, 
    np.dtype, 
    np.dtypes.Float64DType, 
    np.dtypes.UInt32DType,
    np.core.multiarray._reconstruct,
    np.ndarray,
    _codecs.encode, 
])

VER = 1
N_FOLD = 0

os.environ['WANDB_PROJECT'] = f'fixed-clf-bert-fold{N_FOLD}'
os.environ['WANDB_NOTES'] = f'fixed-clf-bert-fold{N_FOLD}'
os.environ['WANDB_NAME'] = f'fixed-clf-bert-fold{N_FOLD}'

df = pd.read_parquet(
    'train_texts.parquet',
    columns=['variantid_1', 'variantid_2', 'group_id', 'is_double']
)

df = df.sort_values(by=['variantid_1', 'variantid_2'])
df = df.sample(len(df), random_state=42)

sgkf = StratifiedGroupKFold(n_splits=5)

fold_mapping = {
    '0': {
        'train_idxs': [],
        'val_idxs': [],
    },
    '1': {
        'train_idxs': [],
        'val_idxs': [],
    },
    '2': {
        'train_idxs': [],
        'val_idxs': [],
    },
    '3': {
        'train_idxs': [],
        'val_idxs': [],
    },
    '4': {
        'train_idxs': [],
        'val_idxs': [],
    },
}

for fold, (train_idx, val_idx) in enumerate(sgkf.split(df, df['is_double'], groups=df['group_id'])):
    fold_mapping[str(fold)]['train_idxs'] = train_idx
    fold_mapping[str(fold)]['val_idxs'] = val_idx

dataset = Dataset.from_parquet('avito-for-dl-train-rev.parquet')
dataset = dataset.rename_column('is_double', 'label')

train_dataset = dataset.select(fold_mapping[str(N_FOLD)]['train_idxs'].tolist())
eval_dataset = dataset.select(fold_mapping[str(N_FOLD)]['val_idxs'].tolist())

len(train_dataset), len(eval_dataset)

train_dataset[0]

@dataclass
class Config:
    output_dir: str = 'output'
    checkpoint: str = 'DeepPavlov/rubert-base-cased'
    num_labels: int = 2
    max_length: int = 512
    optim_type: str = 'adamw_torch'
    per_device_train_batch_size: int = 32
    gradient_accumulation_steps: int = 1
    per_device_eval_batch_size: int = 32
    n_epochs: int = 1
    lr: float = 2e-5
    warmup_ratio: int = 0.03
    
config = Config()

training_args = TrainingArguments(
    output_dir=f'output-{VER}',
    overwrite_output_dir=True,
    report_to='wandb',
    num_train_epochs=config.n_epochs,
    per_device_train_batch_size=config.per_device_train_batch_size,
    gradient_accumulation_steps=config.gradient_accumulation_steps,
    # per_device_eval_batch_size=config.per_device_eval_batch_size,
    logging_steps=1,
    # eval_strategy='epoch',
    save_strategy='steps',
    save_steps=500,
    save_total_limit=4,
    optim=config.optim_type,
    learning_rate=config.lr,
    warmup_ratio=config.warmup_ratio,
    remove_unused_columns=False,
    lr_scheduler_type='cosine',  # 'cosine' or 'linear' or 'constant' (default is 'linear')
)

tokenizer = AutoTokenizer.from_pretrained(config.checkpoint)

model = AutoModelForSequenceClassification.from_pretrained(
    config.checkpoint,
    num_labels=config.num_labels,
    device_map='cuda:0',
    trust_remote_code=True
)

from sklearn.utils.class_weight import compute_class_weight

class_weights = compute_class_weight(
    class_weight='balanced',
    classes=np.unique(df['is_double']),
    y=df['is_double']
)

def collate_fn(batch):
    texts = [item['product_row'] for item in batch]
    labels = torch.tensor([item['label'] for item in batch])
    
    result = tokenizer(
        texts, 
        padding=True, 
        truncation=True, 
        max_length=config.max_length,
        return_tensors="pt"
    )
    
    result["labels"] = labels
    
    return result

class WeightedCETrainer(Trainer):
    def __init__(self, *args, class_weights=None, **kwargs):
        super().__init__(*args, **kwargs)
        if class_weights is not None:
            self.class_weights = torch.tensor(class_weights, dtype=torch.float)

    def compute_loss(self, model, inputs, num_items_in_batch=None, return_outputs=False):
        labels = inputs.get('labels')
        outputs = model(**inputs)
        logits = outputs.get('logits')
        
        loss_fct = torch.nn.CrossEntropyLoss(weight=self.class_weights.to(self.model.device))
        loss = loss_fct(logits.view(-1, self.model.config.num_labels), labels.view(-1))
        return (loss, outputs) if return_outputs else loss

def compute_metrics(eval_preds: EvalPrediction) -> dict:
    y_true = eval_preds.label_ids
    y_pred = eval_preds.predictions[:, 1]
    
    precision, recall, _ = precision_recall_curve(y_true, y_pred)
    prauc = auc(recall, precision)

    return {
        'prauc': prauc
    }

import time
from transformers import TrainerCallback

class TimeLimitCallback(TrainerCallback):
    def __init__(self, time_limit_hours):
        self.time_limit_seconds = time_limit_hours * 3600
        self.start_time = None

    def on_train_begin(self, args, state, control, **kwargs):
        self.start_time = time.time()

    def on_step_end(self, args, state, control, **kwargs):
        elapsed_time = time.time() - self.start_time
        if elapsed_time >= self.time_limit_seconds:
            control.should_training_stop = True

trainer = WeightedCETrainer(
    args=training_args, 
    model=model,
    tokenizer=tokenizer,
    train_dataset=train_dataset,
    # eval_dataset=eval_dataset,
    # compute_metrics=compute_metrics,
    data_collator=collate_fn,
    class_weights=class_weights,
    # callbacks=[TimeLimitCallback(time_limit_hours=11.5)]
)

trainer.train('/kaggle/input/rubert-folds0/output-1/checkpoint-23495')

test_df = pd.read_parquet('avito-for-dl-test-rev.parquet')
test_df['label'] = 0
test_df.loc[228, 'label'] = 1
test_df.loc[1337, 'label'] = 1
test_dataset = Dataset.from_pandas(test_df)

training_args.per_device_eval_batch_size = config.per_device_eval_batch_size
training_args.eval_strategy = 'epoch'
trainer.eval_dataset = test_dataset
trainer.compute_metrics = compute_metrics

output = trainer.predict(test_dataset=test_dataset)

predictions = output.predictions
metrics = output.metrics

import joblib

joblib.dump(predictions, 'predictions_rev.joblib')
joblib.dump(metrics, 'metrics_rev.joblib')

# final_output_dir = f'{config.output_dir}/final_fold{N_FOLD}'
# model.save(final_output_dir)
