==> notebooks/31_rouge_for_texts.ipynb <==
import joblib
import pandas as pd
import numpy as np

train = pd.read_parquet('train_texts.parquet')

train = train.sort_values(by=['variantid_1', 'variantid_2'])
train = train.sample(len(train), random_state=42)

test = pd.read_parquet('test_texts.parquet')

from rouge_metric import PyRouge

metric = PyRouge(
    rouge_n=(1, 2, 3, 4), 
    rouge_l=False, 
    # rouge_w=True, 
    # rouge_w_weight=1.2, 
    rouge_s=True, rouge_su=True, skip_gap=4
)
metric_w = PyRouge(
    rouge_n=(), 
    rouge_l=False, 
    rouge_w=True, 
    # rouge_w_weight=1.2, 
    # rouge_s=True, rouge_su=True, skip_gap=4
)

from razdel import tokenize, sentenize 
from functools import lru_cache

@lru_cache(None)
def s(text):
    return [i.text for i in sentenize(text)]

@lru_cache(None)
def t(text):
    return [i.text for i in tokenize(text)]

metric_w.evaluate(['Диван лофт'], [['Диван книжка']], sentencizer=s, tokenizer=t)

row = train[train['is_double'] == 1].iloc[2]

metric.evaluate(
        [row['name_1'] + '\n' + row['description_1']], 
        [[row['name_2'] + '\n' + row['description_2']]], sentencizer=s, tokenizer=t),\
metric.evaluate(
        [row['name_2'] + '\n' + row['description_2']], 
        [[row['name_1'] + '\n' + row['description_1']]], sentencizer=s, tokenizer=t)

def get_rogues(row):
    r1 = metric.evaluate(
        ['\n'.join([row['name_1'], row['description_1']])], 
        [['\n'.join([row['name_2'], row['description_2']])]], 
        sentencizer=s, tokenizer=t)
    # r2 = metric.evaluate(
    #     [row['name_2'] + '\n' + row['description_2']], 
    #     [[row['name_1'] + '\n' + row['description_1']]], sentencizer=s, tokenizer=t)
    # print(desc_r1)
    # print(desc_r2)
    # rw = metric_w.evaluate(
    #     [row['name_1'] + '\n' + row['description_1']], 
    #     [[row['name_2'] + '\n' + row['description_2']]], sentencizer=s, tokenizer=t)
    # rw2 = metric_w.evaluate(
    #     [row['name_2'] + '\n' + row['description_2']], 
    #     [[row['name_1'] + '\n' + row['description_1']]], sentencizer=s, tokenizer=t)
    
    # return pd.Series([
    return pd.Series([
        # (r1[k]['f']+r2[k]['f'])/2 for k in r1.keys()
        r1[k]['f'] for k in r1.keys()
    # ] + [
    #     (rw['rouge-w-1.2']['f']+rw2['rouge-w-1.2']['f'])/2
    ])

get_rogues(train.iloc[2])

a = train[train['is_double'] == 1].iloc[2]
%timeit get_rogues(a)

from tqdm import tqdm
tqdm.pandas()

rogues = train.progress_apply(get_rogues, axis=1)

rogues_test = test.progress_apply(get_rogues, axis=1)

colnames = ['rouge_' + i for i in ['1', '2', '3', '4', 's4', 'su4']]

rogues.columns = colnames
rogues_test.columns = colnames

train_with_rouge = pd.concat([train[['variantid_1', 'variantid_2']], rogues], axis=1)
test_with_rouge = pd.concat([test[['variantid_1', 'variantid_2']], rogues_test], axis=1)

test_with_rouge

train_with_rouge.to_csv('train_rouge.csv')
test_with_rouge.to_csv('test_rouge.csv')
