==> notebooks/36_final_blending_with_probas_align.ipynb <==
!pip install lightautoml -q

import joblib
import pandas as pd
import numpy as np
from tqdm import tqdm

test = pd.read_parquet(
    'test_df.parquet',
    columns=[
        'variantid_1', 'variantid_2', 
        'category_level_1_1', 'category_level_1_2', 
        'category_level_2_1', 'category_level_2_2', 
        'category_level_3_1', 'category_level_3_2', 
        'category_level_4_1', 'category_level_4_2', 
    ]
)

test

test['unique_cat_1'] = test['category_level_1_1'] + '_' + test['category_level_1_2']
test['unique_cat_2'] = test['category_level_2_1'] + '_' + test['category_level_2_2']
test['unique_cat_3'] = test['category_level_3_1'] + '_' + test['category_level_3_2']
test['unique_cat_4'] = test['category_level_4_1'] + '_' + test['category_level_4_2']

def prauc_metric(y_true, y_pred, sample_weight, **kwargs):
    pass

probas = joblib.load('27-05-avito-lama-pizda/test_pred (4).joblib').data.squeeze()
# probas = joblib.load('19-05-avito-add-rouge-cb/ensemble_predictions_cb.joblib')

test['all_probability'] = probas
test['cat1_probability'] = probas
# test['cat2_probability'] = probas

cat1_mapping = {
    'Электроника': 'el',
    'Хобби и отдых': 'hob',
    'Животные': 'zhiv',
    'Личные вещи': 'lyc',
    'Для дома и дачи': 'dom',
    'Готовый бизнес и оборудование': 'got',
    'Транспорт': 'trans'
}

cat2_mapping = {
    'Коллекционирование': 'kolekc',
    'Оргтехника и расходники': 'orgtech',
    'Ремонт и строительство': 'remont',
    'Телефоны': 'telef',
    'Детская одежда и обувь': 'detod',
    'Одежда, обувь, аксессуары': 'odezda',
    'Оборудование для бизнеса': 'oborud',
    'Аудио и видео': 'audio',
    'Мебель и интерьер': 'mebel',
    'Запчасти и аксессуары': 'zapch',
} # only where cat_size > 40000

len(cat1_mapping)#, len(cat2_mapping)

categories1 = test['unique_cat_1'].unique()
categories1 = [cat.split('_')[0] for cat in categories1 if cat.split('_')[0] == cat.split('_')[1]]
categories1 = [x for x in categories1 if x in cat1_mapping.keys()]

progress_bar = tqdm(categories1)
for category in progress_bar:
    progress_bar.set_description(f"{category=}")

    cat_probas1 = joblib.load(f'3-05-avito-rouge-cat1-10folds/rouge_cat1_10folds/{cat1_mapping[category]}.joblib')
    cat_probas2 = joblib.load(f'24-05-avito-cat1lamas-rouge-10folds/cat1lamas_rouge_10folds/{cat1_mapping[category]}.joblib').data.squeeze()
    # cat_probas = cat_probas1 * 0.7 + cat_probas2 * 0.3
    # cat_probas = cat_probas1

    index = test[test['unique_cat_1'] == f'{category}_{category}'].index
    test.loc[index, 'cat1_probability_1'] = cat_probas1
    test.loc[index, 'cat1_probability_2'] = cat_probas2

# categories2 = test['unique_cat_2'].unique()
# categories2 = [cat.split('_')[0] for cat in categories2 if cat.split('_')[0] == cat.split('_')[1]]
# categories2 = [x for x in categories2 if x in cat2_mapping.keys()]

# progress_bar = tqdm(categories2)
# for category in progress_bar:
#     progress_bar.set_description(f"{category=}")

#     cat_probas = joblib.load(f'/kaggle/input/19-05-avito-cat2-cb-preds/{cat2_mapping[category]}.joblib')

#     index = test[test['unique_cat_2'] == f'{category}_{category}'].index
#     test.loc[index, 'cat2_probability'] = cat_probas

# test['probability'] = test['all_probability']
# test['probability'] = (test['all_probability'] * 0.5 + test['cat1_probability'] * 0.5)
# test['probability'] = (test['all_probability'] * 0.4 + test['cat1_probability'] * 0.4 + test['cat2_probability'] * 0.2)
test

test.rename(columns={
    'variantid_1': 'base_id',
    'variantid_2': 'cand_id',
}, inplace=True)

# temp = pd.concat([test[['base_id', 'cand_id']], test[['base_id', 'cand_id']]], ignore_index=True)

# normalized_pairs = test[['base_id', 'cand_id']].apply(lambda row: tuple(sorted([row['base_id'], row['cand_id']])), axis=1)
# test['pair_key'] = normalized_pairs
# mean_probs = test.groupby('pair_key')['probability'].mean().reset_index()
# mean_probs = mean_probs.rename(columns={'probability': 'mean_probability'})
# test_new = test.merge(mean_probs, on='pair_key', how='left')
# test['probability'] = test_new['mean_probability']
# test_new = test_new.drop(columns=['pair_key', 'mean_probability'])
# test_new

# test[['base_id', 'cand_id']].apply(lambda x: )
# test.iterrows()[:40]
# test[(test['base_id'] in test['cand_id'].tolist()) | (test['cand_id'] in test['base_id'].tolist())]
# test

# test['probability'].plot(kind='hist')
probable_probs = test.groupby('category_level_1_1')[['all_probability', 'cat1_probability_1', 'cat1_probability_2']].mean()

test['fixed_cat1_probability_1'] = test['cat1_probability_1']
test['fixed_cat1_probability_2'] = test['cat1_probability_2']

for model in ['cat1_probability_1', 'cat1_probability_2']:
    for cat in probable_probs.index:
        # target_prob = 
        p1_before = probable_probs.loc[cat]['all_probability']
        p1_after = probable_probs.loc[cat][model]
        p0_before = 1 - p1_before
        p0_after = 1 - p1_after
        p1_actual = test[test['category_level_1_1'] == cat][model]
        p1_adj = (
            p1_actual * p1_before * p0_after) / (
            p1_actual * p1_before * p0_after + (1 - p1_actual) * p0_before * p1_after
        )
        test.loc[test['category_level_1_1'] == cat, 'fixed_' + model] = p1_adj
        # print(p1_actual, p1_adj)
        # print(p1_actual.mean(), p1_adj.mean())
        # p1_actual.plot(kind='hist')
        # p1_adj.plot(kind='hist')
        # display(probable_probs)
        
        # print(target_prob)
        # display(test[test['category_level_1_1'] == cat])


display(probable_probs)
test.groupby('category_level_1_1')[['all_probability', 'fixed_cat1_probability_1', 'fixed_cat1_probability_2']].mean()

test['probability'] = test['all_probability'] * 0.5 + (test['fixed_cat1_probability_1'] * 0.7 + test['fixed_cat1_probability_2'] * 0.3) * 0.5
test['probability'] = test['probability'].fillna(test['all_probability']).clip(0, 1)

# test[test['category_level_1_1'] == 'Животные']['probability'].plot(kind='hist')
test.loc[test['category_level_1_1'] == 'Животные', 'probability'] *= 1.3
test['probability'] = test['probability'].clip(0, 1)
# test[test['category_level_1_1'] == 'Животные']['probability'].plot(kind='hist')
# test[test['category_level_1_1'] == 'Электроника']['probability'].plot(kind='hist')
test.loc[test['category_level_1_1'] == 'Электроника', 'probability'] *= 1.3
test['probability'] = test['probability'].clip(0, 1)
# test[test['category_level_1_1'] == 'Электроника']['probability'].plot(kind='hist')
# test[test['category_level_1_1'] == 'Транспорт']['probability'].plot(kind='hist')
test.loc[test['category_level_1_1'] == 'Транспорт', 'probability'] *= 1.2
test['probability'] = test['probability'].clip(0, 1)
# test[test['category_level_1_1'] == 'Транспорт']['probability'].plot(kind='hist')

test = test[['base_id', 'cand_id', 'probability']]
test

# test.to_csv('nonlin_prob_align_to_bigllama_0_5baselama_3seeds_0_5((0_3cat1lamas_0_7cat1cbs)_10f).csv', index=False)
test.to_csv('z1.3_el1.3_tr1.2_nonlin_prob_align_to_bigllama_0_5baselama_3seeds_0_5((0_3cat1lamas_0_7cat1cbs)_10f).csv', index=False)
